pollkb.sh rewrite
New approach. The udev rule will trigger and run the script for each
device associated with the keyboard, which may or may not be an input
device with a /dev/hidraw* entry, therefore there is no need to loop
over anything in the script itself.

The script gets passed the sysfs device path of the currently detected
device, and all it does is trying to look up an associated hidraw* path.
If it exists and points to an existing chardev with similar name, we
found a candidate and "cat" it for one second to kick the keyboard into
action as a workaround.